requirements.txt
click==8.0.1
PyYAML==5.4.1
requests==2.25.1
retry==0.9.2
